add missing imports + move common functions
