+++ magpie/management/service/service_views.py
@@ -1,6 +1,9 @@
 from management.resource.resource_utils import create_resource
+from service_formats import *
 from service_utils import *
 from api_requests import *
+from common import str2bool
+from register import sync_services_phoenix
 from models import resource_tree_service
 from pyramid.view import view_config
